refactor: moved weekly code to modular files
