+++ components/RecurringTasks/streaks/patterns/weekly/types.ts
@@ -1,0 +1,7 @@
+import { DateRange, WeekDay } from '../../types';
+
+export interface WeeklyTarget {
+  date: Date;
+  allowedRange: DateRange;
+  weekday: WeekDay;
+}
