+++ components/RecurringTasks/streaks/patterns/weekly/utils.ts
@@ -1,0 +1,20 @@
+// Add weekday name mapping
+export const WEEKDAY_NAMES: Record<string, string> = {
+  'sun': 'SUNDAY',
+  'sunday': 'SUNDAY',
+  'mon': 'MONDAY',
+  'monday': 'MONDAY',
+  'tue': 'TUESDAY',
+  'tues': 'TUESDAY',
+  'tuesday': 'TUESDAY',
+  'wed': 'WEDNESDAY',
+  'wednesday': 'WEDNESDAY',
+  'thu': 'THURSDAY',
+  'thur': 'THURSDAY',
+  'thurs': 'THURSDAY',
+  'thursday': 'THURSDAY',
+  'fri': 'FRIDAY',
+  'friday': 'FRIDAY',
+  'sat': 'SATURDAY',
+  'saturday': 'SATURDAY'
+};
